fix: clear remaining 138 ruff issues incl. real bugs
ruff check api auto_client_acquisition core integrations now passes.
Genuine bugs fixed alongside the style cleanup:

- approval_center: missing `ValidationError` import (broke the rogue-field
  422 path) — pre-existing test now passes.
- leads: `/discover/web` called undefined `_search_with_chain` — import added.
- isolated_pg_event_store: `global` omitted three names, causing a
  referenced-before-assignment error in the worker bootstrap.
- admin_tenants: removed an unused `timezone` import shadowed by a field.

S105/S107 hits were reviewed — all false positives (token-type labels,
env-var names, enum members); no real secret found.

+++ api/routers/proof_pack_governed.py
@@ -89,7 +89,7 @@ def proof_generate(engagement_id: str, body: ProofGenerateBody = Body(default_fa
         )
     filled_s = sum(1 for k in PROOF_SECTION_KEYS if body.sections.get(k) not in (None, "", [], {}))
     filled_c = sum(1 for k in COMPONENT_KEYS if body.components.get(k) not in (None, "", [], {}))
-    proof_score = int(round((filled_s / len(PROOF_SECTION_KEYS)) * 70 + (filled_c / len(COMPONENT_KEYS)) * 30))
+    proof_score = round((filled_s / len(PROOF_SECTION_KEYS)) * 70 + (filled_c / len(COMPONENT_KEYS)) * 30)
     tier = "case_candidate" if proof_score >= 85 else ("sales_support" if proof_score >= 70 else ("internal_learning" if proof_score >= 55 else "weak"))
     snap = merge_pipeline_stage(
         engagement_id,
